+++ src/view/uniprotkb/components/__tests__/CatalyticActivityView.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import { CatalyticActivityView } from '../CatalyticActivityView';
+import CatalyticActivityView from '../CatalyticActivityView';
 import data from './modelData.json';
 
 configure({ adapter: new Adapter() });
