+++ src/view/uniprotkb/components/__tests__/FeaturesView.spec.tsx
