+++ src/view/uniprotkb/components/__tests__/GeneNamesView.spec.tsx
