+++ src/view/uniprotkb/components/__tests__/KeywordView.spec.tsx
