+++ src/view/uniprotkb/components/__tests__/NameView.spec.tsx
