+++ src/view/uniprotkb/components/__tests__/ProteinNamesView.spec.tsx
