+++ src/view/uniprotkb/components/__tests__/ProteinOverview.spec.tsx
