+++ src/view/uniprotkb/components/__tests__/SequenceView.spec.tsx
