+++ src/view/uniprotkb/components/__tests__/SimpleView.spec.tsx
